requirements.txt
Jinja2==3.1.1
Flask==2.1.1
python-dotenv==0.20.0
Flask-JWT-Extended==4.3.1
mysql-connector-python==8.0.28
PyJWT==2.3.0
pyOpenSSL==22.0.0
PySocks==1.7.1
boto3==1.22.0
botocore==1.25.0
awscli==1.23.0
